refactor(profile): profile_tab_page.dart 추가 분해 (1755→843줄, -52%)
좌측/우측 패널 빌더 블록을 추가 part 파일로 분리.
이전 분해(2628→1755)에서 더 나아가 843줄까지 축소.

신설:
- lib/widgets/profile/profile_left_panel.dart (~718줄)
  · _buildProfileLeftPanel, _buildProfileContentTabs, _profileContentTabButton
  · _buildProfileContentPanel, _buildProfileNotesTabBody
  · _buildProfileVersesTabBody, _buildProfilePrayerTabBody
  · _profileTabSectionHeader, _profileInlineTextButton, _buildProfileTabMessage
  · _buildProfileNotePreviewCard, _buildProfileSavedVersePreviewCard
  · _openProfileNotePreview, _buildCurrentUserAvatar
- lib/widgets/profile/profile_right_panel.dart (~208줄)
  · _buildProfileRightPanel, _profileTestamentToggle
  · _profileMiniActionButton, _profileShareIdChip

부모 파일: 1755 → 843줄 (-52%)
총 profile 분해: 2628 → 843줄 (-68%)

flutter analyze: 0 issues
flutter test: 183 tests pass
